pymongo/requirements.txt
numpy
pytest
debugpy
pymongo[srv]
python-dotenv
matplotlib
pandas
statsmodels